+++ server/server_net.h
@@ -25,5 +25,5 @@ along with Czech_mate.  If not, see <http://www.gnu.org/licenses/>.
 	extern void net_init();
 	extern void net_engine();
 	extern void net_close();
-	extern int net_loop_settings;
+	extern int_ net_loop_settings;
 #endif
